Delete original Api.ts and replace with new Api file
